CHG: api now uses writer & reader + ADD: connectionManager and factory

+++ include/ConnectionConfiguration.hpp
@@ -13,12 +13,12 @@ namespace ghost
 		 
 		// getters of connection configuration parameters
 		int getConnectionId() const;
-		int getThreadPoolSize() const;
+		size_t getThreadPoolSize() const;
 		bool isOperationBlocking() const;
 
 		// setters of connection configuration parameters
 		void setConnectionId(int id);
-		void setThreadPoolSize(int size);
+		void setThreadPoolSize(size_t size);
 		void setOperationBlocking(bool value);
 	};
 }
